feat(sponsor): only show logo on homepage
